move tests to related project, added more tests
