+++ NUnitTests/NUnitTests.csproj
@@ -14,6 +14,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\BusinessLogic\BusinessLogic.csproj" />
+    <ProjectReference Include="..\SharedForTests\SharedForTests.csproj" />
   </ItemGroup>
 
 </Project>
