+++ UnitTestFrameworksComparison.sln
@@ -3,13 +3,15 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30104.148
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CoreFunctionalityTests", "UnitTestsFrameworkComparison\CoreFunctionalityTests.csproj", "{54EAB250-0D90-4C0E-92CF-EDA4453B9141}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MSTestAndFluentTests", "UnitTestsFrameworkComparison\MSTestAndFluentTests.csproj", "{54EAB250-0D90-4C0E-92CF-EDA4453B9141}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BusinessLogic", "BusinessLogic\BusinessLogic.csproj", "{8165C28D-D74F-4B9C-8D6C-D8AFE21EE748}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NUnitTests", "NUnitTests\NUnitTests.csproj", "{1127742E-C7BA-4CDB-93CC-C8E9DC8BC63A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "XUnitTests", "XUnitTests\XUnitTests.csproj", "{B2270832-47DB-47EE-A9BB-700EFFBAC963}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharedForTests", "SharedForTests\SharedForTests.csproj", "{E6D5118F-29F4-4C18-BD80-A7C63E2ED42E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,6 +35,10 @@ Global
 		{B2270832-47DB-47EE-A9BB-700EFFBAC963}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B2270832-47DB-47EE-A9BB-700EFFBAC963}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B2270832-47DB-47EE-A9BB-700EFFBAC963}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E6D5118F-29F4-4C18-BD80-A7C63E2ED42E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E6D5118F-29F4-4C18-BD80-A7C63E2ED42E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E6D5118F-29F4-4C18-BD80-A7C63E2ED42E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E6D5118F-29F4-4C18-BD80-A7C63E2ED42E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
