+++ UnitTestsFrameworkComparison/Events/EventExamples.cs
@@ -1,12 +1,7 @@
 ﻿using BusinessLogic.PurchaseOrderModule;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using NUnit;
-using CoreFunctionality.TestHelper;
-using BusinessLogic.Interfaces;
+using SharedForTests.PurchaseOrderModule;
 
 namespace CoreFunctionality.Events
 {
@@ -41,22 +36,6 @@ namespace CoreFunctionality.Events
                 // assert
                 eventSource.Should().Raise("ItemAdded");
             }
-        }
-
-        [TestMethod]
-        public void NUnit_Adding_Item_Should_Raise_Event()
-        {
-            // arrange
-            PurchaseItem purchaseItem = new PurchaseItem().CreateRandom();
-           
-            bool wasRaised = false;
-            _purchaseOrder.ItemAdded += (_, s) => { wasRaised = true; };
-
-            // act
-            _purchaseOrder.AddPurchaseItem(purchaseItem);
-
-            // assert
-            NUnit.Framework.Assert.IsTrue(wasRaised);
         }
     }
 }
