+++ UnitTestsFrameworkComparison/MSTestAndFluentTests.csproj
@@ -21,6 +21,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\BusinessLogic\BusinessLogic.csproj" />
+    <ProjectReference Include="..\SharedForTests\SharedForTests.csproj" />
   </ItemGroup>
 
 </Project>
